Moved ToolConsumerProfile class to Lti\v2
