+++ LtiLibrary.AspNetCore.Tests/Profiles/ToolConsumerProfileController.cs
@@ -2,8 +2,8 @@
 using System.Threading.Tasks;
 using LtiLibrary.AspNetCore.Profiles;
 using LtiLibrary.NetCore.Common;
+using LtiLibrary.NetCore.Lti.v2;
 using LtiLibrary.NetCore.Lti2;
-using LtiLibrary.NetCore.Profiles;
 using NodaTime;
 
 namespace LtiLibrary.AspNetCore.Tests.Profiles
