+++ LtiLibrary.NetCore.Tests/ToolConsumerProfileTests.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using LtiLibrary.NetCore.Tests.SimpleHelpers;
 using LtiLibrary.NetCore.Common;
+using LtiLibrary.NetCore.Lti.v2;
 using LtiLibrary.NetCore.Lti2;
-using LtiLibrary.NetCore.Profiles;
 using NodaTime;
 using Xunit;
